Move Resource SSM config to a separate package
