+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/configuration/state_machine/resource/StateMachinePersistenceConfig.java
@@ -1,4 +1,4 @@
-package eu.bbmri.eric.csit.service.negotiator.configuration.state_machine;
+package eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.resource;
 
 import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationResourceEvent;
 import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationResourceState;
